chore: adjusts component exports
