+++ packages/plugin-multi-tenant/src/components/TenantSelector/index.tsx
@@ -1,46 +1,47 @@
-import type { OptionObject, ServerProps } from 'payload'
+'use client'
+import type { ReactSelectOption } from '@payloadcms/ui'
 
-import { cookies as getCookies } from 'next/headers.js'
+import { SelectInput } from '@payloadcms/ui'
 import React from 'react'
 
-import type { UserWithTenantsField } from '../../types.js'
+import './index.scss'
+import { SELECT_ALL } from '../../constants.js'
+import { useTenantSelection } from '../../providers/TenantSelectionProvider/index.client.js'
 
-import { TenantSelectorClient } from './index.client.js'
+export const TenantSelector = () => {
+  const { options, selectedTenantID, setTenant } = useTenantSelection()
 
-type Args = {
-  tenantsCollectionSlug: string
-  useAsTitle: string
-  user?: UserWithTenantsField
-} & ServerProps
+  const handleChange = React.useCallback(
+    (option: ReactSelectOption | ReactSelectOption[]) => {
+      if (option && 'value' in option) {
+        setTenant({ id: option.value as string, refresh: true })
+      } else {
+        setTenant({ id: undefined, refresh: true })
+      }
+    },
+    [setTenant],
+  )
 
-export const TenantSelector = async ({
-  payload,
-  tenantsCollectionSlug,
-  useAsTitle,
-  user,
-}: Args) => {
-  const { docs: userTenants } = await payload.find({
-    collection: tenantsCollectionSlug,
-    depth: 0,
-    limit: 1000,
-    overrideAccess: false,
-    sort: useAsTitle,
-    user,
-  })
-
-  const tenantOptions: OptionObject[] = userTenants.map((doc) => ({
-    label: String(doc[useAsTitle]),
-    value: String(doc.id),
-  }))
-
-  const cookies = await getCookies()
-  let selectedTenant = tenantOptions.find(
-    (tenant) => tenant.value === cookies.get('payload-tenant')?.value,
-  )?.value
-
-  if (!selectedTenant && userTenants?.[0]) {
-    selectedTenant = String(userTenants[0].id)
+  if (options.length <= 1) {
+    return null
   }
 
-  return <TenantSelectorClient initialValue={selectedTenant} options={tenantOptions} />
+  return (
+    <div className="tenant-selector">
+      <SelectInput
+        label="Tenant"
+        name="setTenant"
+        onChange={handleChange}
+        options={options}
+        path="setTenant"
+        value={
+          selectedTenantID
+            ? selectedTenantID === SELECT_ALL
+              ? undefined
+              : String(selectedTenantID)
+            : undefined
+        }
+      />
+    </div>
+  )
 }
