+++ packages/plugin-multi-tenant/src/exports/client.ts
@@ -1,2 +1,2 @@
 export { TenantField } from '../components/TenantField/index.client.js'
-export { TenantSelectionProvider } from '../providers/TenantSelectionProvider/index.js'
+export { TenantSelector } from '../components/TenantSelector/index.js'
