+++ packages/plugin-multi-tenant/src/exports/rsc.ts
@@ -1,2 +1,2 @@
 export { GlobalViewRedirect } from '../components/GlobalViewRedirect/index.js'
-export { TenantSelector } from '../components/TenantSelector/index.js'
+export { TenantSelectionProvider } from '../providers/TenantSelectionProvider/index.js'
